Remove static/ directory and update frontend deployment docs
Followup on commit 2991254 which removed the management commands
necessary for serving the frontend via Django. The supported workflow
is now to deploy the frontend separately and configure it to talk
to this Django app. If desired, both can be served by an Nginx proxy
(which serves static files for the frontend and proxies /api requests
to Django).